requirements.txt
torch
torchvision
pillow
git+https://github.com/modestyachts/ImageNetV2_pytorch
